refactor(daemon): move the memory subdomain into src/memory/
Pure file moves: the eight memory files leave src/agents/ for a dedicated
src/memory/ directory, with the two runtime-facing ones under
src/memory/runtime/. No symbol renames, no barrel, no boundary changes —
only the moved files' own relative imports, ~35 importers across src and
test, and the doc paths that pointed at the old locations.

  agents/memory-provider.ts        -> memory/provider.ts
  agents/memory.ts                 -> memory/store.ts
  agents/memory-fs.ts              -> memory/fs.ts
  agents/memory-recall.ts          -> memory/recall.ts
  agents/memory-distiller.ts       -> memory/distill.ts
  agents/native-memory.ts          -> memory/runtime/native.ts
  agents/runtime-memory.ts         -> memory/runtime/capabilities.ts
  agents/managed-distill-outbox.ts -> memory/managed-distill-outbox.ts
