+++ packages/daemon/src/evaluation/runner.ts
@@ -18,7 +18,7 @@ import { detectSandbox } from '../acp/sandbox.js'
 import { effectiveRunInSandbox } from '../acp/runtime-launch.js'
 import { agentChildEnv } from '../agents/agent-env.js'
 import { selectAgent } from '../agents/load-agents.js'
-import { memoryProviderFor } from '../agents/memory-provider.js'
+import { memoryProviderFor } from '../memory/provider.js'
 import { loadConfig } from '../config/load-config.js'
 import { Daemon, type DaemonEvaluationOptions, type DaemonEvaluationTurnResult } from '../daemon.js'
 import { composeRuntimeLaunch } from '../runtimes/launch-policy.js'
