+++ packages/daemon/src/k8s/runtime-plane.ts
@@ -10,7 +10,7 @@ import { ShimFileSink } from '../shim/channels.js'
 import { ShimWorkspaceFiles } from '../shim/workspace-files-channel.js'
 import { ShimMemoryFs } from '../shim/memory-fs-channel.js'
 import type { WorkspaceFiles } from '../workspace/workspace-files.js'
-import type { MemoryFs } from '../agents/memory-fs.js'
+import type { MemoryFs } from '../memory/fs.js'
 import { DEFAULT_SHIM_LISTEN_PORT, DEFAULT_SHIM_WORKSPACE_ROOT } from '../shim/protocol.js'
 import type { TunnelName } from '../shim/tunnel.js'
 import type { ShimSession } from '../shim/session.js'
