+++ packages/daemon/test/session-manager.test.ts
@@ -4,9 +4,9 @@ import { tmpdir } from 'node:os'
 import { join } from 'node:path'
 import { LocalStore, sessionKey, transcriptChannelKey } from '../src/store/local-store.js'
 import { SessionManager, isStandingContextTitleEcho } from '../src/session/session-manager.js'
-import { createManagedMemoryProvider } from '../src/agents/memory-provider.js'
-import { writeMemoryFile, MEMORY_INDEX, MAX_INDEX_INJECT_BYTES } from '../src/agents/memory.js'
-import { LocalMemoryFs } from '../src/agents/memory-fs.js'
+import { createManagedMemoryProvider } from '../src/memory/provider.js'
+import { writeMemoryFile, MEMORY_INDEX, MAX_INDEX_INJECT_BYTES } from '../src/memory/store.js'
+import { LocalMemoryFs } from '../src/memory/fs.js'
 import type { Agent } from '../src/agents/agent-schema.js'
 import type { NormalizedMessage } from '../src/messages/normalized.js'
